Add bounds constraints for expansion / parsing
Add the ability to place bounds on which territories can be returned,
i.e only EU territories, even if asked to expand WW or are passed a
longer list - basically returns the intersection of the bounds and the
input.

This is really useful when you have a hard set of constraints on what
territories something can be used in, but more specific restrictions
that may sometimes be too liberal.